NamedPipe::write: Remove waitUntilWritten option
Windows-only: The option to return from the NamedPipe::write function
before the write operation has finished creates a dangling pointer to
the OVERLAPPED structure that is created on the Stack since a pointer to
it is passed to the Windows API call WriteFile.
However, if NamedPipe::write returns before WriteFile has finished, this
API function now deals with a dangling pointer, which is no good.

Since the waitUntilWritten option was only supported on Windows anyway,
the option was simply removed.