Use busy "sleep" in certain situations
If the overall timeout of e.g. a call to read_blocking is very short,
the solution of ea8fcf7a347ad7414df86cb4d6dc7b028b2d8a75 is no longer
sufficient. It'll work as long as the overall timeout is large enough to
not notice up to 15ms delay (and makes sure that delay doesn't
accumulate), but it breaks down, when the overall timeout is e.g. 1ms.

For these cases, we actually need a (somewhat) precise control over how
long the program sleeps, which is why we have to resort to busy waiting
in these cases.